feat: logged in user update and delete contact

+++ controllers/contactController.js
@@ -3,16 +3,16 @@ const Contact = require("../models/contactModel");
 
 //@desc get all contacts
 //@route GET /api/contacts
-//@access public
+//@access private
 
 const getContacts = asyncHandler(async (req, res) => {
-  const contacts = await Contact.find();
+  const contacts = await Contact.find({ user_id: req.user.id });
   res.status(200).json(contacts);
 });
 
 //@desc create  new contact
 //@route POST /api/contacts
-//@access public
+//@access private
 const createContact = asyncHandler(async (req, res) => {
   console.log(req.body);
   const { name, email, phone } = req.body;
@@ -24,13 +24,14 @@ const createContact = asyncHandler(async (req, res) => {
     name,
     email,
     phone,
+    user_id: req.user.id,
   });
   res.status(200).json(contact);
 });
 
 //@desc get contact
 //@route GET /api/contacts/:id
-//@access public
+//@access private
 const getContact = asyncHandler(async (req, res) => {
   const contact = await Contact.findById(req.params.id);
   if (!contact) {
@@ -42,13 +43,19 @@ const getContact = asyncHandler(async (req, res) => {
 });
 //@desc update contact
 //@route get /api/contacts/
-//@access public
+//@access private
 const updateContact = asyncHandler(async (req, res) => {
   const contact = await Contact.findById(req.params.id);
   if (!contact) {
     res.status(404);
     throw new Error("contact not found");
   }
+
+  if (contact.user_id.toString() !== req.user.id) {
+    res.status(403);
+    throw new Error("user don't have permission to update this contact");
+  }
+
   const updatedContact = await Contact.findByIdAndUpdate(
     req.params.id,
     req.body,
@@ -59,13 +66,18 @@ const updateContact = asyncHandler(async (req, res) => {
 });
 //@desc get contact
 //@route get /api/contacts/:id
-//@access public
+//@access private
 const deleteContact = asyncHandler(async (req, res) => {
   const contact = await Contact.findById(req.params.id);
   if (!contact) {
     res.status(404);
     throw new Error("Contact not found");
   }
+  if (contact.user_id.toString() !== req.user.id) {
+    res.status(403);
+    throw new Error("user don't have permission to update this contact");
+  }
+
   // yalnızca bir kişiyi silmek için
   await Contact.findByIdAndDelete(req.params.id);
   res.status(200).json(contact);
